requirements.txt
Flask==1.1.1
Jinja==1.2
setuptools==46.1.1
